add a new BGZF filetype - copied from BgzfFileType
The goal is to provide an alternative reader that is aware of
the underlying block and record structures of BGZF files.  This
should allow corrupted BAM files to be read, then re-written with
relatively low loss rates.

This commit is a stub and is simply a copied and renamed version
of BgzfFileType with no more or less functionality.

BgzfFileTypeRecovery has not been tested other than it compiles
and links.