eliminate dependency on openssl for md5
history: openssl has a modern implementation of the
reference version of md5 checksums.  However, the
support issues simply in trying to have users load
open ssl development libraries on all platforms is
not worth it.

Here we pull the original reference implementation
from the IETF, converted to C++, hidden as statics
inside a .h file, and use it inline.